Update examples.py
The pin mode for the blink method was off. Took out the port configuration in the main because the library can usually find it on it's own, and was causing issues when I was manually specifying.
Currently using for a project, so I'm sure I'll have more commits for you eventually.